Fix "See also" references
1. There are two types of "x" references in NamesList.txt,
one with parentheses and names, the second only stating the
referenced code point. Support for the second type was added.

2. The regexp for the first type was incorrect for code
points with more than 4 hex digits. The additional digits
were eaten by the .* regexp preceeding the capture.

For example, the reference to U+1F4B2 in the details list for
character U+0024 DOLLAR SIGN was recognized as F4B2.

3. Added missing range check; the current data file format
only supports 16 bit references.