Add ability to set window color scheme manually
It's still optional, currently it will attempt to find KF6ConfigWidgets
and the related feature will be enabled if KF6ConfigWidgets can be found.

This can be handy for non-Linux users since their OS might not yet offer
a native dark color scheme, or they might want to switch to a color scheme
which is different from the one the OS currently used.

This patch also set the Qt theme to breeze due to the current Windows11
style have some minor issues. We might also want to add an option to set
the application style like what Kdenlive did.